Moves all profile related functionality into the mapstory_profile app.  This includes models, views, urls, forms and tests.
