+++ mapstory/settings/base.py
@@ -179,7 +179,7 @@ TEMPLATES = [
 #
 # Authentication Settings
 #
-ACCOUNT_ADAPTER = os.environ.get('ACCOUNT_ADAPTER', 'mapstory.views.MapStoryAccountAdapter')
+ACCOUNT_ADAPTER = os.environ.get('ACCOUNT_ADAPTER', 'mapstory.mapstory_profile.views.MapStoryAccountAdapter')
 ACCOUNT_FORMS = {'signup': 'mapstory.forms.CustomSignupForm'}
 ACCOUNT_EXTRA_PROFILE_FORM = os.environ.get('ACCOUNT_EXTRA_PROFILE_FORM', 'EditMapstoryProfileForm')
